Decouple Public API and impl details
