+++ packages/cache/src/index.ts
@@ -1,1 +1,18 @@
 export { buildCache } from './cache.js';
+export type {
+  Cache,
+  CacheTransaction,
+  LiveCacheTransaction,
+  CommittingTransaction,
+  CacheEntry,
+  CacheEntryState,
+  CacheKeyValue,
+  EntityMergeStrategy,
+  RevisionMergeStrategy,
+  CachedEntityRevision,
+  ExpirationPolicy,
+  CacheOptions,
+  DefaultRegistry,
+  CacheDebugAPIs,
+  CacheTransactionDebugAPIs
+} from './api.js'
